websocket phase change notification
